requirements.txt
# Core Python packages
pandas>=1.5.3
numpy>=1.24.3
scipy>=1.10.1
matplotlib>=3.7.1
seaborn>=0.12.2
plotly>=5.16.1

# Database
psycopg2-binary>=2.9.8
SQLAlchemy>=2.0.20

# API
fastapi>=0.111.0
uvicorn[standard]>=0.30.0

# Model persistence
joblib>=1.3.2

# Validation
pydantic>=1.10.13

# Jupyter
notebook>=6.5.4
ipykernel>=6.25.0

# Dashboard / Visualization
dash>=2.11.0
dash-bootstrap-components>=1.4.1

# Data Cleaning / Utility
openpyxl>=3.1.2
xlrd>=2.0.1
requests>=2.31.0

# Optional: ML/EDA
scikit-learn>=1.2.2

# Testing (optional)
pytest>=7.4.2
httpx>=0.27.0

# Environment management
python-dotenv>=1.0.0